- Fixed test suite to not abort with fatal errors, if no DSN is provided.
git-svn-id: https://svn.apache.org/repos/asf/incubator/zetacomponents/trunk@978705 13f79535-47bb-0310-9956-ffa450edef68

+++ Database/tests/sqlabstraction/expression_test.php
@@ -68,6 +68,8 @@ class ezcQueryExpressionTest extends ezcTestCase
 
     protected function tearDown()
     {
+        if ( $this->db === null ) return;
+
         $this->db->exec( 'DROP TABLE query_test' );
     }
