+++ Database/tests/sqlabstraction/query_select_test_impl.php
@@ -62,6 +62,8 @@ class ezcQuerySelectTestImpl extends ezcTestCase
 
     protected function tearDown()
     {
+        if ( $this->db === null ) return;
+
         $this->db->exec( 'DROP TABLE query_test' );
         $this->db->exec( 'DROP TABLE query_test2' );
     }
